+++ easytier-gui/src-tauri/.cargo/config.toml
@@ -1,0 +1,4 @@
+[build]
+target = "x86_64-pc-windows-msvc"
+
+[target]
