+++ easytier-gui/src-tauri/gen/android/app/.gitignore
@@ -1,0 +1,6 @@
+/src/main/java/com/kkrainbow/easyiter/generated
+/src/main/jniLibs/**/*.so
+/src/main/assets/tauri.conf.json
+/tauri.build.gradle.kts
+/proguard-tauri.pro
+/tauri.properties
