+++ easytier-gui/src-tauri/gen/android/app/src/main/java/com/kkrainbow/easyiter/MainActivity.kt
@@ -1,0 +1,3 @@
+package com.kkrainbow.easyiter
+
+class MainActivity : TauriActivity()
