+++ easytier-gui/src-tauri/gen/android/app/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">easytier-gui</string>
+    <string name="main_activity_title">easytier-gui</string>
+</resources>
