+++ easytier-gui/src-tauri/gen/android/settings.gradle
@@ -1,0 +1,3 @@
+include ':app'
+
+apply from: 'tauri.settings.gradle'
