+++ easytier/src/common/ifcfg.rs
@@ -12,25 +12,37 @@ pub trait IfConfiguerTrait: Send + Sync {
         name: &str,
         address: Ipv4Addr,
         cidr_prefix: u8,
-    ) -> Result<(), Error>;
+    ) -> Result<(), Error> {
+        Ok(())
+    }
     async fn remove_ipv4_route(
         &self,
         name: &str,
         address: Ipv4Addr,
         cidr_prefix: u8,
-    ) -> Result<(), Error>;
+    ) -> Result<(), Error> {
+        Ok(())
+    }
     async fn add_ipv4_ip(
         &self,
         name: &str,
         address: Ipv4Addr,
         cidr_prefix: u8,
-    ) -> Result<(), Error>;
-    async fn set_link_status(&self, name: &str, up: bool) -> Result<(), Error>;
-    async fn remove_ip(&self, name: &str, ip: Option<Ipv4Addr>) -> Result<(), Error>;
+    ) -> Result<(), Error> {
+        Ok(())
+    }
+    async fn set_link_status(&self, name: &str, up: bool) -> Result<(), Error> {
+        Ok(())
+    }
+    async fn remove_ip(&self, name: &str, ip: Option<Ipv4Addr>) -> Result<(), Error> {
+        Ok(())
+    }
     async fn wait_interface_show(&self, _name: &str) -> Result<(), Error> {
         return Ok(());
     }
-    async fn set_mtu(&self, _name: &str, _mtu: u32) -> Result<(), Error>;
+    async fn set_mtu(&self, _name: &str, _mtu: u32) -> Result<(), Error> {
+        Ok(())
+    }
 }
 
 fn cidr_to_subnet_mask(prefix_length: u8) -> Ipv4Addr {
@@ -379,6 +391,10 @@ impl IfConfiguerTrait for WindowsIfConfiger {
     }
 }
 
+pub struct DummyIfConfiger {}
+#[async_trait]
+impl IfConfiguerTrait for DummyIfConfiger {}
+
 #[cfg(target_os = "macos")]
 pub type IfConfiger = MacIfConfiger;
 
@@ -387,3 +403,6 @@ pub type IfConfiger = LinuxIfConfiger;
 
 #[cfg(target_os = "windows")]
 pub type IfConfiger = WindowsIfConfiger;
+
+#[cfg(not(any(target_os = "macos", target_os = "linux", target_os = "windows")))]
+pub type IfConfiger = DummyIfConfiger;
